requirements.txt
pandas
geopandas
sqlalchemy
psycopg2-binary 
tqdm
chardet
geoalchemy2
dotenv
matplotlib
seaborn 
mapclassify
contextily
esda 
libpysal 
splot